requirements.txt
pandas
numpy
MetaTrader5
scikit-learn
hmmlearn
matplotlib
python-dotenv
